wb-bot-backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0 
django-cors-headers==4.7.0